backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
h11==0.14.0
idna==3.10
passlib==1.7.4
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.45.3
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.3.0
uv==0.5.30
uvicorn==0.34.0
Jinja2
sqlalchemy
psycopg2-binary==2.9.9